Add support for ext-session-lock-v1 protocol
gtk4-layer-shell supports the ext-session-lock starting with version
1.1.0. Even though both protocols use the same shared library, they
produce different GIR files, which means that we need to generate
separate sys and api crates for them.

This commit adds the gtk4-session-lock and gtk4-session-lock-sys crates
as well as a simple example analogous to the one in gtk4-layer-shell.
The GIR file was generated from the source of version 1.1.0 of
gtk4-layer-shell.

This commit does not add the files generated by gir yet to keep the diff
small. The files are generated by CI automatically in the following
commits.