CI: Commit updates of submodules if any of the code changed
We updated the submodules but then never commited their changes. This lead to failed "Build" actions when only the newest version understood new elements in .gir files. We now also only commit the changes if some of the code changed and not when only the version.txt files changed. This should lead to less commit noise in the future